Refs #3361. Don't try to load nonexistant icons..
I noticed this when working on the preview generation. ChannelInfo::LoadIcon() was attempting to load icons even if the iconpath was empty. This results in a lot of unneeded chatter between the backend and frontend. I only noticed this because I thought a storage group helper function I had added was failing to find valid files, but it turned out to be an existing problem.

git-svn-id: http://svn.mythtv.org/trunk@15027 4b91aca1-2a1f-0410-a73a-a78f2df169e8