Refs #3361. Improves Preview Generation.
 * Implements a priority queue for preview generation and limits the total number of in-flight preview generation threads.
 * Runs preview generator in a separate process so that an ffmpeg crash can't take down the backend or frontend we're using.

Currently the number of preview generation threads is limited to two by the 'previewGeneratorMaxRunning' variable. I haven't yet implemented the third todo, adding user control over various aspects of the preview generation processing.

Note: This also extends the screen shot code to work with remote backends, and fixes a race condition that caused this to crash sometimes. However unless both the backend and frontend are updated to this revision a remote 'screen shot' will just return a copy of the preview image.

git-svn-id: http://svn.mythtv.org/trunk@15029 4b91aca1-2a1f-0410-a73a-a78f2df169e8