Closes #4327. Refs #3980. UK EIT fixup by john {at} pullan {dot} org
Does a bunch of cleanups to the EIT fixups for the UK, which make
it more correct and fix a few niggly issues that people have noticed.

Now handles CSI properly as well as some cases where the ... joining
code was incorrectly propagating the whole title/subtitle/desc into
the title.

Only one small change from the original patch, which makes sure the
airdate being used as a title/description delimiter still works.



git-svn-id: http://svn.mythtv.org/trunk@15178 4b91aca1-2a1f-0410-a73a-a78f2df169e8