Changes the way MythThemedDialogs are painted. Previously a context change could leave portions of a container in an area that wasn't redrawn (switching to a context with a container that is larger than the original container and back, for example).
Previously, containers started with a rect of the background
(MythThemedDialog background) and were drawn over that.
This lead to "rough" edges where spots of background would
boarder a container that didn't paint every pixel of it's area. This
is no longer the case, every container in an invalidated rect is drawn
to the same pixmap, allowing overlaps.


git-svn-id: http://svn.mythtv.org/trunk@15200 4b91aca1-2a1f-0410-a73a-a78f2df169e8