Do for regions what was done in [15200].
DrawRegion() will no longer be called in response to a requestRegionUpdate.
Instead Draw() is called, allowing laying to work as it does elsewhere
(regions no longer exist because they always do, each container
is only permitted to draw in the invalidated rect as of [15200]).

UIListTreeType was the only control to do this, if you see strange
painting behavior open a ticket.


git-svn-id: http://svn.mythtv.org/trunk@15216 4b91aca1-2a1f-0410-a73a-a78f2df169e8