Drop the VideoGalleryAspectRatio setting from mythvideo.
The value of this setting is questionable, in 99% of case ignoring the aspect ratio of the original image would produce undesirable results and especially with movie posters. The setting wouldn't survive the mythui port anyway since the existing gallery widget will be binned in favour of something much better and with greater theming potential.



git-svn-id: http://svn.mythtv.org/trunk@15253 4b91aca1-2a1f-0410-a73a-a78f2df169e8