This slightly convoluted change to the http/xml GetPreviewImage function attempts to speed up preview generation for mythweb and support for preserving the aspect ratio of a rescaled preview image.
First, if a default preview image exists then rescale that to the
required size instead of generating a new one from the video. This is
several seconds faster on my backend for just a single image.

Second, if you supply a Height but no Width, or a Width but no Height,
then scale the image to that dimension whilst preserving aspect ratio.

The PreviewPixmapOffset is now honoured if SecsIn is not defined or is
equal to zero. If you supply a SecsIn argument > 0 or "!= 
PreviewPixmapOffset" then a new image will be generated.

There is a change to the filename format.
Old:
filename.WxHxS.png
New:
filename.S{.WxH}.png

Closes #3217



git-svn-id: http://svn.mythtv.org/trunk@15328 4b91aca1-2a1f-0410-a73a-a78f2df169e8