Enhanced the category search to allow searching by genre if that information is available.  For example, this weekend's NFL games could previously only be found under the "Sports event" category.  They can now be found under "Sports event", "Football", "Playoff sports" or any combination of two of them.  Likewise, tonight's episode of CSI can be found under "Crime drama", "Action", "Adventure", "Suspense" or any combination.
Besides making it easier to find things of known interest, browsing
the list of genre combinations for odding pairings can lead to
potential nuggets that might otherwise be missed.  For example, who'd
have thought there would be a program which matched both "Home
improvement" and "Religious" genres?

Because the list of genre combinations can be very large, the
MythListBox widget has been modified to make it easier to quickly
navigate through the choose view popup dialog.  In short, press 0
through 9 to move to Nx10 percent in the list and press
PREVVIEW/NEXTVIEW to move to the previous/next major grouping.  That
change was committed separately since it is not specific to category
searches.

If genre information is not available, the catefory search works like
before using only the category information found in the program table.


git-svn-id: http://svn.mythtv.org/trunk@15411 4b91aca1-2a1f-0410-a73a-a78f2df169e8