Add a few ffmpeg options to mythburn.py when transcoding from nuv to mpeg2.
You will also need to create a custom encoding profile to take advantage of
these new encoding parameters

I haven't tested this because all my files are encoded to mpeg2 and it doesn't
make sense to convert them to nuv if they are going to be burned to DVD if you
ask me :-)

Patch by thrantastic (at) gmail.com. Closes #4539.

git-svn-id: http://svn.mythtv.org/trunk@15612 4b91aca1-2a1f-0410-a73a-a78f2df169e8