Rework some internals in the StorageGroup class to make it easier to add some special groups that are being worked on.
This patch also makes it so that you can no longer create new scheduled
recordings to go into the LiveTV Storage Group.  If you want to put a regular
recording in the same dir as LiveTV recordings, then you need to create a new
Storage Group pointed at the same directory or add that directory to an
existing Storage Group.

Closes #4556 using patch by Michael T. Dean.



git-svn-id: http://svn.mythtv.org/trunk@15699 4b91aca1-2a1f-0410-a73a-a78f2df169e8