Merge MythHorizListButton with MythListButton and introduce the concept of layouts. Added a third grid format layout as well.
This turns MythListButton into a verstile widget that can be used for
everything from simple lists, to the gallery views in
MythGallery/MythVideo. Even better, the layout used is themeable
meaning that all sorts of different things can be done with themes.

The "buttonlist" definition now includes a <layout> element which
accepts the values vertical, horizontal and grid.



git-svn-id: http://svn.mythtv.org/trunk@15721 4b91aca1-2a1f-0410-a73a-a78f2df169e8