A little clean up to MythArchive's native archive creation.
Actually create the ISO image when creating a native archive and the user has
that option checked. It was only working if you specified a directory to save
to before and not when a DVD format was chosen.

Catch when burning a DVD or creating an ISO image fails so it doesn't report
the archive was created OK when it only partly worked.

git-svn-id: http://svn.mythtv.org/trunk@15740 4b91aca1-2a1f-0410-a73a-a78f2df169e8