Fixes #4498. Makes ChannelBase try a little harder to find a valid channel on a restricted recorder w/patch from Shane.
When a DVB recorder is tuned to a particular multiplex and another recorder referring to the same device tries to tune to another frequency we instead try to find a channel on the same multiplex. The channel browser and channel up/down were already restricted to valid channels, but before this patch was applied you could force the recorder to another channel through direct entry. The backend would actually prevent you from putting the backend into an inconsistent state, but the frontend would go on merrily believing that it would be getting a signal on the new channel any moment now..

git-svn-id: http://svn.mythtv.org/trunk@15760 4b91aca1-2a1f-0410-a73a-a78f2df169e8