Add a menu option to allow the user to tell Myth to forget that it recorded a particular recording so that it can be rerecorded even if it is a duplicate.
The 'forgotten' program also no longer counts in the maxepisodes count for
the scheduled recording so that it may be rerecorded.

This is useful when you have a bad recording of a program and want to give
Myth a chance to record a new copy without having to delete the old one first.

Closes #3917 using patch by Tony Lill.



git-svn-id: http://svn.mythtv.org/trunk@15808 4b91aca1-2a1f-0410-a73a-a78f2df169e8