Move the weather grabbing scripts into subdirectories according to the source they grab from. This is purely to make it easier to manage the scripts down the line but is also necessary for that same reason.
NOTE you will need to rerun mythweather setup twice, the first time so
that it deletes the old screens and a second time to add the re-add
them. Apologies for this, but better this change is made now than
after 0.21 is released where it will cause much more disruption.

Thanks to Joe Ripley for the original patch which made this possible.

Refs #3722



git-svn-id: http://svn.mythtv.org/trunk@15908 4b91aca1-2a1f-0410-a73a-a78f2df169e8