References #4791
Fixes a problem where novideos_info had no area, resulting in an empty pixmap,
which apparently cannot be drawn on. An error is now output when playwait or
novideos_info has an invalid rect.

Also removes novideos_info from the "selected" window as it should have
been impossible to ever see.

Thanks to hitch_123 [a] yahoo for finding this bug.


git-svn-id: http://svn.mythtv.org/trunk@16312 4b91aca1-2a1f-0410-a73a-a78f2df169e8