Merges [16357] and closes #2662
This is a work around for when libav doesnt return a valid DTS
after a call to av_seek_frame().  This allows us to seek around
in MKV/MOV/MP4 files. If at some point libav changes to return
some info we can remove it but for now this does the job.




git-svn-id: http://svn.mythtv.org/trunk@16362 4b91aca1-2a1f-0410-a73a-a78f2df169e8