Partial port of mythgallery to mythui.
This converts the main screen of mythgallery to mythui. It works, but is not complete and there are some remaining issues. e.g. Folders don't have the folder image and thumbnail images are not rescaled to fit after rotating. Both these issues are trivial and will be fixed shortly. I'm committing this now to allow Janne to port the plugins to QT4.



git-svn-id: http://svn.mythtv.org/trunk@16597 4b91aca1-2a1f-0410-a73a-a78f2df169e8