Added the missing support for overriding the audio sampling rate in recording profiles on a per card basis.  mythtv-setup previously allowed the user to configure capturecard.audioratelimit, but it was not actually used anywhere.  If set to a value other than "None", it is now used to force audio recording at the specified rate.
Use of this setting is intended for users with a mix of capture cards
that do not all support the same sampling rates.  For example I have
one card which only supports 32 kHz.  Previously, I had to set all
recording profiles to the noticably lower quality 32 kHz rate because
of one card.  Now, I can force just that one card to always use 32 kHz
and the other cards can use higher rates.

Because some users might have inadvertently changed the setting and
left it that way when it wasn't used, this change resets it back to
"None" for all cards.  This will prevent unexpected changes in
configurations that previously worked.


git-svn-id: http://svn.mythtv.org/trunk@16747 4b91aca1-2a1f-0410-a73a-a78f2df169e8