Fixes #3504. Make sure we always find the EPG start channel.
This adds a GuideGrid::FindChannel() method which finds the EPG start channel. Before multirec we only stored the chanid for the currently shown channel, even if we had two or more chanid represented by a single row in the EPG. Now that we keep the chanid info we can use it to more reliably locate the EPG start channel.

git-svn-id: http://svn.mythtv.org/trunk@16882 4b91aca1-2a1f-0410-a73a-a78f2df169e8