Refs #5107. Refs #2514. Preliminary support for video playback window resizing.
Mostly committing this now so I can use the new monitor dimension variables to handle #2514 properly.

This is disabled by default. If you want to try it, run mythfrontend with "-O UseFixedWindowSize=0".

This does not do the right thing if you enter the EPG and may not play well with display resolution switching based on video dimensions.

git-svn-id: http://svn.mythtv.org/trunk@16893 4b91aca1-2a1f-0410-a73a-a78f2df169e8