References #4862, #5220
This imports a modified version of the AC3 CRC check introduced in
ffmpeg's revision 11379.

This should eliminate chirp/pop issues when the AC3 data is corrupt.

Thanks to Tom Dexter for this patch. My modifications were only to make
it as close to the current ffmpeg code as possible (without importing the
new CRC table stuff).


git-svn-id: http://svn.mythtv.org/trunk@17083 4b91aca1-2a1f-0410-a73a-a78f2df169e8