Fix a few MythWeather problems.
Fix retrieving the list of scripts caused by the change to Qt4 - 
QDir::entryInfoList() now returns a QList<QFileInfo> and not a
QPtrList<QFileInfo> *.

Start to convert some Qt3 compatibility classes to Qt4 friendly ones.

Fix some SourceManager segfaults when not using background fetching. Both
the Weather screen and the ScreenSetup screen will now create and delete
their own copy of the SourceManager if they are passed a NULL pointer for
the SourceManager.

git-svn-id: http://svn.mythtv.org/trunk@17097 4b91aca1-2a1f-0410-a73a-a78f2df169e8