Adds what I'm calling 'Centre Hold' scrolling to mythlistbutton. This is where the selected item remains static in the centre of the visible list area and the list items move instead. The behaviour is similar to that of the category list in "Watch Recordings".
Adds a new element to the listbutton, <scrollstyle> which accepts values of "free", the current behaviour and "center", the new behaviour. The intention is to allow further styles to be added later.

This was tested with the vertical layout and the horizontal layout should work too, but it doesn't yet work with the grid layout and fixing that isn't a high priority at the moment. There are several reasons for adding this option now, but the one which will interest most people is to take us one step closer to a "cover flow" effect.

Refs #12


git-svn-id: http://svn.mythtv.org/trunk@17174 4b91aca1-2a1f-0410-a73a-a78f2df169e8